+++ pytest_splinter/__init__.py
@@ -1,2 +1,2 @@
 """pytest-splinter package."""
-__version__ = '1.3.4'
+__version__ = '1.3.5'