+++ tests/mocked_browser/test_browser_overrides.py
@@ -5,11 +5,7 @@ import mock
 import pytest
 
 
-def test_wait_for_condition_default(
-    browser,
-    splinter_browser_load_condition,
-    splinter_browser_load_timeout,
-):
+def test_wait_for_condition_default(browser, splinter_browser_load_condition, splinter_browser_load_timeout):
     """Test that by default wait_until is successful."""
     browser.wait_for_condition(
         splinter_browser_load_condition,
@@ -18,10 +14,7 @@ def test_wait_for_condition_default(
     assert True
 
 
-def test_wait_for_condition_timeout(
-    browser,
-    monkeypatch,
-):
+def test_wait_for_condition_timeout(browser, monkeypatch):
     """Check timeouts."""
     ticks = iter([1, 2, 15])